Fixed reinterpret_cast. Result is 64-bit integer.

+++ code/mathematical_algorithms/src/fast_inverse_sqrt/fast_inverse_sqrt.cpp
@@ -12,7 +12,7 @@
 double fastInverseSqrt(double x)
 {
     float xhalf = 0.5f * x;
-    int32_t i = reinterpret_cast<int32_t>(x);              // get bits for floating value
+    int64_t i = reinterpret_cast<int64_t>(x);              // get bits for floating value
     i = 0x5f375a86 - (i >> 1);      // gives initial guess y0
     x = reinterpret_cast<float>(i);                // convert bits back to float
     x = x * (1.5f - xhalf * x * x); // Newton step, repeating increases accuracy
